Add Authorization to Swagger
After logging in successfully using Swagger, you can retrieve your JWT. Scroll to the top and hit the 'Authorize' button in the top right. Enter in the following format 'Bearer {Token}', replacing {Token} with the JWT and hitting the 'Authorize' button will authorize the user to access the endpoints they have permission to execute.